pyproject.toml
[project]
name = "loopcat"
version = "0.1.0"
description = "Catalog WAV files from Boss RC-300 looper pedal with AI-powered audio analysis"
readme = "plan.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "xxhash>=3.0.0",
    "textual>=0.50.0",
    "sounddevice>=0.4.6",
    "simple-term-menu>=1.6.0",
]

[project.optional-dependencies]
midi = [
    "mido>=1.3.0",
    "python-rtmidi>=1.5.0",
]

[project.scripts]
loopcat = "loopcat.cli:app"
gridcat = "gridcat.__main__:main"
fadercat = "fadercat.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loopcat", "src/cat_common", "src/gridcat", "src/fadercat"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
